IL - changed carryingCapacity to be type long

+++ src/main/java/edu/ucsb/cs156/happiercows/entities/Commons.java
@@ -31,7 +31,7 @@ public class Commons
   private LocalDateTime startingDate;
   private LocalDateTime endingDate;
   private double degradationRate;
-  private int carryingCapacity;
+  private long carryingCapacity;
   private boolean showLeaderboard;
 
   @ManyToMany(fetch = FetchType.EAGER, cascade = {CascadeType.PERSIST,CascadeType.REMOVE})
